requirements.txt
absl-py==1.4.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
configparser==6.0.0
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.12.2
fonttools==4.42.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
packaging==23.1
pathtools==0.1.2
Pillow==10.0.0
promise==2.3
protobuf==3.20.1
psutil==5.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
sentry-sdk==1.30.0
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
sympy==1.12
tensorboard-data-server==0.7.1
# torch==2.0.1+cu118
# torchaudio==2.0.2+cu118
torchnet==0.0.4
# torchvision==0.15.2+cu118
tornado==6.3.3
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.7.1
urllib3==1.26.16
visdom==0.2.4
wandb==0.10.21
websocket-client==1.6.2
Werkzeug==2.3.7
zipp==3.16.2
